Allow organisers to always edit regos and make registration/x/status work for other people.
It turns out that if a organiser has "paid" for their invoice then they
can no longer modify peoples registrations.  Make the payment check only
apply to non-organisers.

The /register/status page (really /registration/x/status) was fully
hardcoded to only function for the logged in user.  Change it to work
against a person.  This makes it a nicer experience for organisers
investigating registrations.