Fix bug #339.
http://rt.followtheflow.org/Ticket/Display.html?id=339

The mini-confs have a very different set of requirements for the organizer
than a paper-submitter, in particular all the consents are wrong; and the
view page does not show the target audience.  The problem with the 
consents is that viewing the miniconf proposal shows the submitter as
having consented to a lot of things which they did not consent to.

Thus create a special template for miniconfs and switch to that using
the fragment field in the view_base.mako template.