Allow users to be logged in unactivated
Make the sign up process easier by not requiring users to be activated
to be logged in. Require an activated account for paper submission and
conference registration. Warn on granting privs to in-active account.

For the moment redirect to an info placeholder. Soon to be replaced with
fancy callback stuff.