Changed missing role behaviour
If a role is requested but not in the database the code threw an
exception. I have changed this to throw a NotAuthorizedError.

This makes the failure a graceful rather than terminal one.
The change is being done to make testing easier, it should also help
development allowing code to be written before the database change is
made.

There is no impact on a production site, the code should not run.