Added functional testing for CRUD like controllers
The tests are implemented using a CrudHelper class which the tests
inherit from.

CrudHelper implements the test suite:
 * test_permissions
 * test_new
 * test_index
 * test_view
 * test_edit
 * test_delete

Each function takes as its required parameters: self, app, db_session
Each function also takes a large variety of optional parameters,
these allow the function to be tweaked and customised depending on the
expected output. Common customisations are things like a different form
of page title or a custom target object to test for.

There are three ways these functions can be used by each test file:
 1. No code written, the generic test in inherited and used
 2. Customisation code written, the generic test is called from
    within the test file function of the same name providing customised
    options and typically extra code before or after.
 3. CrudHelper code overwritten, the test file has a function of the
    same name but does not call the CrudHelper function. This allows
    full custom testing but is typically used to disable the test if the
    controller does not support the function.

Improved diagnostic of mako errors
----------------------------------

The best solution is to change the stack trace so you get a nice Mako
based trace rather than the 30 odd full list

This was the prior behaviour provided by pylons.error handle_mako_error

Unfortunately it doesn't work, I spent hours trying to work it out but
it seems that the stack trace is being rewritten again by one of the
third party layers.

So for now we do the stack rewrite - it might start working

We also print out the nice Mako stack trace so we have something to work
with in the mean time.

Testing improvements
--------------------

Make login function reentrant
This simplifies tests by not requiring a logout and state tracking

Bug fixes
---------

In addition to implementing the tests a stack of minor bug fixes have
also been applied as the bugs were uncovered.

This range from the very minor - changing the case of the template title
string to be consistent.

To the relatively major - existing functionality did not compile and was
obviously never tested.

Due to the large number of these changes they were not broken out in to
individually reported issues.