Remove DMCA protected files
These are apparently protected. However, the license on the head
of the javascript is actually MIT. So I'm not sure which parts are
licensed as what but since we've only been given 24hours to remove
the content I'm just going to purge it all.

This may break some themes/pages/elements. But given that zookeepr
is largely unmaintained at this point and that most deployments
write their own theme I'm not too worried.